+++ app/login/page.tsx
@@ -1,7 +1,8 @@
+"use client";
 import React from "react";
 import Head from "next/head";
-import AppLayout from "../components/AppLayout";
-import LoginForm from "../components/LoginForm";
+import AppLayout from "../../components/AppLayout";
+import LoginForm from "../../components/LoginForm";
 
 const Login = () => {
   return (
