+++ app/profile/page.tsx
@@ -1,8 +1,9 @@
+"use client";
 import React from "react";
 import Head from "next/head";
-import AppLayout from "../components/AppLayout";
-import UserProfile from "../components/UserProfile";
-import StoredPost from "../components/StoredPost";
+import AppLayout from "../../components/AppLayout";
+import UserProfile from "../../components/UserProfile";
+import StoredPost from "../../components/StoredPost";
 
 const Profile = () => {
   const storeP = [
