+++ pages/_app.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import Head from "next/head";
-// import "antd/dist/antd.css";
-import wrapper from "../store/configureStore";
+import wrapper from "../../store/configureStore";
 
 interface Prop {
   Component: () => React.JSX.Element;
